+++ src/actions/check_payment_required.js
@@ -1,92 +1,53 @@
 'use strict';
 
-var JsonProducer = require('rc-kinesis').JsonProducer;
-var CwrxRequest = require('../../lib/CwrxRequest');
-var resolveURL = require('url').resolve;
-var ld = require('lodash');
-var moment = require('moment');
-var q = require('q');
+const JsonProducer = require('rc-kinesis').JsonProducer;
+const CwrxRequest = require('../../lib/CwrxRequest');
+const resolveURL = require('url').resolve;
+const ld = require('lodash');
+const moment = require('moment');
 
-module.exports = function checkPaymentRequiredFactory(config) {
-    var watchmanStream = new JsonProducer(config.kinesis.producer.stream, config.kinesis.producer);
-    var request = new CwrxRequest(config.appCreds);
-    var paymentsEndpoint = resolveURL(config.cwrx.api.root, config.cwrx.api.payments.endpoint);
-    var paymentMethodsEndpoints = resolveURL(paymentsEndpoint, 'methods');
-    var orgsEndpoint = resolveURL(config.cwrx.api.root, config.cwrx.api.orgs.endpoint);
+module.exports = function factory(config) {
+    const watchmanStream = new JsonProducer(
+        config.kinesis.producer.stream,
+        config.kinesis.producer
+    );
+    const request = new CwrxRequest(config.appCreds);
+    const paymentsEndpoint = resolveURL(config.cwrx.api.root, config.cwrx.api.payments.endpoint);
+    const paymentMethodsEndpoints = resolveURL(paymentsEndpoint, 'methods');
+    const paymentPlansEndpoint = resolveURL(
+        config.cwrx.api.root,
+        config.cwrx.api.paymentPlans.endpoint
+    );
 
-    return function checkPaymentRequired(event) {
-        var data = event.data;
-        var org = data.org;
-        var paymentPlanStart = org.paymentPlanStart && moment(org.paymentPlanStart);
-        var nextPaymentDate = org.nextPaymentDate && moment(org.nextPaymentDate);
+    return event => Promise.resolve().then(() => {
+        const data = event.data;
+        const org = data.org;
+        const now = moment(data.date);
+        const nextPaymentDate = org.nextPaymentDate && moment(org.nextPaymentDate);
 
-        function needsPayment(lastPayment) {
-            var today = moment(data.date);
-            var nextPaymentDate = lastPayment && moment(lastPayment.createdAt).add(1, 'month');
-
-            return !lastPayment || nextPaymentDate.isSameOrBefore(today, 'day');
+        if (!nextPaymentDate || nextPaymentDate.isAfter(now)) {
+            return undefined;
         }
 
-        function produceRecord() {
-            return request.get({ url: paymentMethodsEndpoints, qs: { org: org.id } })
-                .spread(function getDefaultPaymentMethod(paymentMethods) {
-                    var paymentMethod = ld.find(paymentMethods, { default: true });
+        return Promise.all([
+            request.get({ url: paymentMethodsEndpoints, qs: { org: org.id } }),
+            request.get({ url: `${paymentPlansEndpoint}/${org.paymentPlanId}` })
+        ]).then(ld.unzip).then(ld.spread(ld.spread((paymentMethods, paymentPlan) => {
+            const paymentMethod = ld.find(paymentMethods, { default: true });
 
-                    if (!paymentMethod) {
-                        throw new Error('Org ' + org.id + ' has no payment methods.');
-                    }
-
-                    return watchmanStream.produce({
-                        type: 'paymentRequired',
-                        data: {
-                            org: org,
-                            paymentPlan: config.paymentPlans[org.paymentPlanId],
-                            paymentMethod: paymentMethod,
-                            date: data.date
-                        }
-                    });
-                });
-        }
-
-        function updateOrg(nextPaymentDate) {
-            return request.put({
-                url: orgsEndpoint + '/' + org.id,
-                json: {
-                    nextPaymentDate: nextPaymentDate.format()
-                }
-            });
-        }
-
-        if (nextPaymentDate) {
-            if (nextPaymentDate.isSameOrBefore(moment(data.date))) {
-                return produceRecord();
+            if (!paymentMethod) {
+                throw new Error('Org ' + org.id + ' has no payment methods.');
             }
 
-            return q();
-        }
-
-        if (!(org.paymentPlanId in config.paymentPlans) || !paymentPlanStart) {
-            return q();
-        }
-
-        if (paymentPlanStart.isAfter(moment(data.date), 'day')) {
-            return updateOrg(paymentPlanStart).thenResolve(undefined);
-        }
-
-        return request.get({ url: paymentsEndpoint, qs: { org: org.id } })
-            .spread(function checkPayments(payments) {
-                var today = moment(data.date);
-                var lastPayment = ld.find(payments, { status: 'settled' });
-                var paymentRequired = needsPayment(lastPayment);
-                var nextPaymentDate = paymentRequired ?
-                    today : moment(lastPayment.createdAt).add(1, 'month');
-
-                return updateOrg(nextPaymentDate).then(function() {
-                    if (paymentRequired) {
-                        return produceRecord();
-                    }
-                });
-            })
-            .thenResolve(undefined);
-    };
+            return watchmanStream.produce({
+                type: 'paymentRequired',
+                data: {
+                    org,
+                    paymentPlan,
+                    paymentMethod,
+                    date: data.date
+                }
+            });
+        })));
+    });
 };