+++ src/actions/showcase/apps/auto_increase_budget.js
@@ -1,38 +1,31 @@
 'use strict';
 
-var CwrxRequest = require('../../../../lib/CwrxRequest');
-var resolveURL = require('url').resolve;
-var ld = require('lodash');
-var filter = ld.filter;
-var assign = ld.assign;
-var get = ld.get;
-var floor = ld.floor;
-var q = require('q');
-var Status = require('cwrx/lib/enums').Status;
-var logger = require('cwrx/lib/logger');
-var inspect = require('util').inspect;
+const CwrxRequest = require('../../../../lib/CwrxRequest');
+const resolveURL = require('url').resolve;
+const ld = require('lodash');
+const filter = ld.filter;
+const assign = ld.assign;
+const get = ld.get;
+const floor = ld.floor;
+const q = require('q');
+const Status = require('cwrx/lib/enums').Status;
+const logger = require('cwrx/lib/logger');
+const inspect = require('util').inspect;
 
-module.exports = function autoIncreaseBudgetFactory(config) {
-    var log = logger.getLog();
-    var request = new CwrxRequest(config.appCreds);
-    var campaignsEndpoint = resolveURL(config.cwrx.api.root, config.cwrx.api.campaigns.endpoint);
+module.exports = function factory(config) {
+    const log = logger.getLog();
+    const request = new CwrxRequest(config.appCreds);
+    const campaignsEndpoint = resolveURL(config.cwrx.api.root, config.cwrx.api.campaigns.endpoint);
 
-    return function autoIncreaseBudget(event) { return q().then(function() {
-        var data = event.data;
-        var options = event.options;
-        var transaction = data.transaction;
-        var transactionDescription = (function() {
-            try {
-                return JSON.parse(transaction.description) || {};
-            } catch(error) {
-                throw new Error('"' + transaction.description + '" is not JSON.');
-            }
-        }());
-        var dailyLimit = options.dailyLimit;
-        var paymentPlan = config.paymentPlans[transactionDescription.paymentPlanId];
+    return event => Promise.resolve().then(() => {
+        const data = event.data;
+        const options = event.options;
+        const transaction = data.transaction;
+        const dailyLimit = options.dailyLimit;
+        const paymentPlan = config.paymentPlans[transaction.paymentPlanId];
 
         if (!paymentPlan) {
-            throw new Error('Unknown payment plan id: ' + transactionDescription.paymentPlanId);
+            throw new Error('Unknown payment plan id: ' + transaction.paymentPlanId);
         }
 
         return request.get({
@@ -46,16 +39,16 @@ module.exports = function autoIncreaseBudgetFactory(config) {
                     Status.OutOfBudget, Status.Error
                 ].join(',')
             }
-        }).spread(function increaseBudgets(campaigns) {
-            var appCampaigns = filter(campaigns, { product: { type: 'app' } });
+        }).spread(campaigns => {
+            const appCampaigns = filter(campaigns, { product: { type: 'app' } });
             // Split this transaction between all showcase (app) campaigns
-            var externalImpressionPortion = floor(
+            const externalImpressionPortion = floor(
                 (transaction.amount / appCampaigns.length) * paymentPlan.impressionsPerDollar
             );
-            var budgetPortion = (transaction.amount / appCampaigns.length);
+            const budgetPortion = (transaction.amount / appCampaigns.length);
 
-            return q.all(appCampaigns.map(function increaseBudget(campaign) {
-                var externalCampaign = get(campaign, 'externalCampaigns.beeswax', {});
+            return q.all(appCampaigns.map(campaign => {
+                const externalCampaign = get(campaign, 'externalCampaigns.beeswax', {});
 
                 return request.put({
                     url: campaignsEndpoint + '/' + campaign.id,
@@ -66,8 +59,8 @@ module.exports = function autoIncreaseBudgetFactory(config) {
                             dailyLimit: dailyLimit
                         })
                     })
-                }).spread(function increaseExternalBudget(newCampaign) {
-                    return request.put({
+                }).spread(newCampaign => (
+                    request.put({
                         url: campaignsEndpoint + '/' + campaign.id + '/external/beeswax',
                         json: {
                             budgetImpressions: (externalCampaign.budgetImpressions || 0) +
@@ -76,7 +69,7 @@ module.exports = function autoIncreaseBudgetFactory(config) {
                             budget: null,
                             dailyLimit: null
                         }
-                    }).spread(function logSuccess(newExternalCampaign) {
+                    }).spread(newExternalCampaign => {
                         log.info(
                             'Increased budget of campaign(%1): %2 => %3.',
                             campaign.id, get(campaign, 'pricing.budget', 0),
@@ -85,13 +78,14 @@ module.exports = function autoIncreaseBudgetFactory(config) {
                         log.info(
                             'Increased budget of externalCampaign(%1): %2 => %3.',
                             externalCampaign.externalId,
-                            externalCampaign.budget, newExternalCampaign.budget
+                            externalCampaign.budgetImpressions,
+                            newExternalCampaign.budgetImpressions
                         );
-                    });
-                });
+                    })
+                ));
             }));
         });
-    }).catch(function logError(reason) {
-        return log.error('Failed to increase campaign budget: %1', inspect(reason));
-    }).thenResolve(undefined); };
+    }).catch(reason => (
+        log.error('Failed to increase campaign budget: %1', inspect(reason))
+    )).then(() => undefined);
 };